% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBinomProb.R
\name{calcBinomProb}
\alias{calcBinomProb}
\title{Calculates the binomial probabilities}
\usage{
calcBinomProb(counts, prop, target.LS, target.total.count)
}
\arguments{
\item{counts}{see @param counts of RNAsub}

\item{prop}{see @param prop of RNAsub}

\item{target.LS}{see @param target.LS of RNAsub}

\item{target.total.count}{see @param target.total.count of RNAsub}
}
\value{
a vector of binomial probabilities
}
\description{
Calculates the binomial probabilities
}
