% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputValidation.R
\name{checkInputs}
\alias{checkInputs}
\title{Check the validty of user inputs}
\usage{
checkInputs(counts, prop, target.LS, target.total.count)
}
\arguments{
\item{counts}{see @param counts of RNAsub}

\item{prop}{see @param prop of RNAsub}

\item{target.LS}{see @param target.LS of RNAsub}

\item{target.total.count}{see @param target.total.count of RNAsub}
}
\value{
Error or warning messages if inputs are not the way they are required to be.
}
\description{
Check the validty of user inputs
}
