% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creat.exampl.RNAseq.data.R
\name{creat.exampl.RNAseq.data}
\alias{creat.exampl.RNAseq.data}
\title{Example RNA-seq data}
\usage{
creat.exampl.RNAseq.data(G = 10000, n = 10, disp = 0.8,
  rlt.abund = NULL, L = NULL, L.param = list(log.mean = 15, scale =
  0.1), ...)
}
\arguments{
\item{G}{number of genes}

\item{n}{number of samples}

\item{disp}{over-dispersion parameter for NB distribution}

\item{rlt.abund}{relative abundances}

\item{L}{library sizes}

\item{L.param}{log.normal distribution parameters for library sizes (if L is not NULL)}

\item{...}{other arguments}
}
\value{
a matrix of counts with dimension Gxn
}
\description{
creates a synthetic data from negative binomial distribution
}
\examples{
# default
counts <- creat.exampl.RNAseq.data()
dim(counts)
head(counts[, 1:5])

#custom
counts <- creat.exampl.RNAseq.data(G=1000, n=5, L.param = list(log.mean=10, scale=0.1))
dim(counts)
head(counts[, 1:5])

}
