% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsubPackage.R
\docType{package}
\name{RNAsub-package}
\alias{RNAsub-package}
\title{RNAsub package}
\description{
A simplified method to subsample a given RNA-seq data to a 
desired level of either library size per sample or total count.
}
\author{
Alemu Takele Assefa \email{alemutak@hotmail.com}
}
