% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsub.R
\name{RNAsub}
\alias{RNAsub}
\title{RNA-seq subsampling}
\usage{
RNAsub(counts, prop = NULL, target.LS = NULL,
  target.total.count = NULL, seed = 64312,
  BPPARAM = BiocParallel::SerialParam(), ...)
}
\arguments{
\item{counts}{orginal (unormalized) RNA-seq count matrix (genes in row and samples in column)}

\item{prop}{A binomial probability (0<prop<1). Either a single value to subsample the
total read counts or a vector of dimension equal to the number of  samples to
subsample each sample to a particular library size. For the later case, the new
library size is equal to prop*L, where L is the library sizes in the original
count matrix.}

\item{target.LS}{The desired library size per sample (if prop is not specified).
It can be a single value or a vector of dimension equal to the number of  samples.}

\item{target.total.count}{The desired total number of reads (if prop or target.LS
are not specified). It is a single numeric value.}

\item{seed}{an integer number to control the random subsampling procedure especially
when one wants to reproduce results. See \code{set.seed()}.}

\item{BPPARAM}{An optional BiocParallelParam instance determining the parallel back-end to be 
used during evaluation, or a list of BiocParallelParam instances, to be applied in 
sequence for nested calls to BiocParallel functions.}

\item{...}{additional arguments to pass to internal functions, such as \code{rbinom}}
}
\value{
a matrix of counts with dimension equal to the original data
}
\description{
A simplified method to subsample a given RNA-seq data to a
desired level of either library size per sample or total count. This method
is adopted from the subSeq bioconductor package.
}
\details{
Let X is a Gxn matrix of un-normalized read counts with genes in rows and samples
in columns. In addition, let Y be the required matrix of subsampled read counts with Gxn
dimension. The read counts of gene g in sample i in Y, denoted as y_gi, is assumed to have
a binomial distribution with probability P_gi and trial size x_gi.

(1)	If the user specifies the subsampling proportion (via prob=p) and if it is scalar, then
the probability of binomial distribution, denoted as P_gi, will be p for every count data in X.
This results in a count matrix Y that has a total count of p times the total count in X.
If the user specifies a vector of subsampling proportions, let prob=c(p_1, p_2, …,p_n), with
dimension equal to the number of samples, then the ith sample in Y will have a library size of
p_i time the library size of ith sample in X.

(2)	If the user specifies the desired library size (via target.LS=l) and if it
is scalar, then the probability of binomial distribution for sample i, denoted as P_i,
will be l/L_i, where L_i is the library size of the ith sample in X. This results in a
count matrix Y that has a library size of exactly l for every sample. If a vector of desired
library sizes are provided, i.e. target.LBb=c(l_1, l_2, …, l_n), then the output matrix Y will
have n samples each with different library size equal to their corresponding library
size in target.LS.

(3)	If the user specifies the desired total read-counts (via target.total.count=t), then
the probability of binomial distribution, denoted as P_gi, will be p=t/T for every count data in X,
where T is the total read counts in X (t<=T). This results in a count matrix Y that has a total count
of p times the total count in X.
}
\examples{
\donttest{
# generate artificial RNA-seq data from negative binomial distribution with fixed
counts <- creat.exampl.RNAseq.data() # see \\code{?creat.exampl.RNAseq.data}

# subsample from counts
sub1 <- RNAsub(counts, prop=0.25, seed=642)
colSums(sub1)
sum(sub1)
sum(sub1)/sum(counts)

sub2 <- RNAsub(counts, target.LS=5000, seed=642)
colSums(sub2)

sub3 <- RNAsub(counts, target.total.count=50000, seed=642)
colSums(sub3)
sum(sub3)
}

}
\references{
David Robinson, John D. Storey and with contributions from Andrew J.
Bass (2019). subSeq: Subsampling of high-throughput sequencing count data. R package
version 1.14.0. http://github.com/StoreyLab/subSeq
}
\author{
Alemu Takele Assefa (alemutak@hotmail.com)
}
