% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSample.R
\name{subsample}
\alias{subsample}
\title{subsample}
\usage{
subsample(counts.cc, cc, n.samples, binom.prob, seed, ...)
}
\arguments{
\item{counts.cc}{counts[, cc]}

\item{cc}{column number}

\item{n.samples}{number of samples}

\item{binom.prob}{binomial probability}

\item{seed}{see @param seed in ?RNAsub}

\item{...}{additional arguments for rbinom or sapply}
}
\value{
a vector of subsampled counts
}
\description{
subsample
}
